refactor: #653 PriceCompareUseCase houme-application으로 이동 (AGENTS.md 규약 준수)
